move price endpoint to price route
