+++ app.js
@@ -11,7 +11,8 @@ const usersRouter = require('./routes/users');
 const coinsRouter = require('./routes/coins');
 const testsRouter = require('./routes/test');
 const dataRouter = require('./routes/data');
-const trackRouter = require('./routes/track')
+const trackRouter = require('./routes/track');
+const priceRouter = require('./routes/price');
 
 const app = express();
 
@@ -29,7 +30,8 @@ app.use('/users', usersRouter);
 app.use('/coins', coinsRouter);
 app.use('/test', testsRouter);
 app.use('/data', dataRouter);
-app.use('/track', trackRouter)
+app.use('/track', trackRouter);
+app.use('/price', priceRouter);
 
 
 // catch 404 and forward to error handler
